test(hooks): inline repo-scoped dist build lock in symlink tests
Drop the shared coordination module (SLizard P1) and keep build serialization checkout-local under .deft-scratch/; revert unrelated bin/slice reader hooks.
