+++ packages/cli/src/bin.test.ts
@@ -3,7 +3,6 @@ import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
 import { engineInfo } from "@deftai/directive-core";
 import { describe, expect, it } from "vitest";
-import { waitForCliDistBuildIdle } from "./cli-dist-test-coordination.js";
 
 const VERSION_LINE = `@deftai/directive (engine: @deftai/directive-core@${engineInfo().version})\n`;
 
@@ -11,13 +10,11 @@ describe("dist/bin.js entrypoint", () => {
   const binPath = join(dirname(fileURLToPath(import.meta.url)), "../dist/bin.js");
 
   it("prints the engine banner for --version", () => {
-    waitForCliDistBuildIdle();
     const out = execFileSync("node", [binPath, "--version"], { encoding: "utf8" });
     expect(out).toBe(VERSION_LINE);
   });
 
   it("prints the same banner for -V", () => {
-    waitForCliDistBuildIdle();
     const out = execFileSync("node", [binPath, "-V"], { encoding: "utf8" });
     expect(out).toBe(VERSION_LINE);
   });
